random weights to links

+++ cmd/graphModel.go
@@ -5,15 +5,15 @@ type Node struct {
 	ID      string  `json:"id"`
 	Group   int     `json:"group"`
 	Layer   int     `json:"layer"`
-	Value   float32 `json:"value"`
+	Value   float64 `json:"value"`
 	ActivFn string  `json:"activFn"`
 }
 
 // Link structure (golang), source and target will be IDs
 type Link struct {
-	Source string `json:"source"`
-	Target string `json:"target"`
-	Weight int    `json:"weight"`
+	Source string  `json:"source"`
+	Target string  `json:"target"`
+	Weight float64 `json:"weight"`
 }
 
 // GraphData holds both nodes and links
